app/back/requirements.txt
grpc-interceptor ~= 0.11.0
grpcio == 1.59.3 
grpcio-tools == 1.59.3